+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/listing/reviewer/UrlReviewer.java
@@ -8,6 +8,7 @@ import gov.healthit.chpl.domain.CertificationResult;
 import gov.healthit.chpl.domain.CertifiedProductSearchDetails;
 import gov.healthit.chpl.permissions.ResourcePermissions;
 import gov.healthit.chpl.util.ErrorMessageUtil;
+import gov.healthit.chpl.util.Util;
 import gov.healthit.chpl.util.ValidationUtils;
 
 /**
@@ -18,9 +19,13 @@ import gov.healthit.chpl.util.ValidationUtils;
 @Component("urlReviewer")
 public class UrlReviewer extends PermissionBasedReviewer {
 
+    private ValidationUtils validationUtils;
+
     @Autowired
-    public UrlReviewer(ErrorMessageUtil msgUtil, ResourcePermissions resourcePermissions) {
+    public UrlReviewer(ValidationUtils validationUtils, ErrorMessageUtil msgUtil,
+            ResourcePermissions resourcePermissions) {
         super(msgUtil, resourcePermissions);
+        this.validationUtils = validationUtils;
     }
 
     @Override
@@ -47,10 +52,10 @@ public class UrlReviewer extends PermissionBasedReviewer {
     private void addListingErrorIfNotValid(final CertifiedProductSearchDetails listing,
             final String input, final String fieldName) {
         if (!StringUtils.isEmpty(input)) {
-            if (ValidationUtils.hasNewline(input)) {
+            if (validationUtils.hasNewline(input)) {
                 listing.getErrorMessages().add(
                         msgUtil.getMessage("listing.invalidUrlFound", fieldName));
-            } else if (!ValidationUtils.isWellFormedUrl(input)) {
+            } else if (!validationUtils.isWellFormedUrl(input)) {
                 listing.getErrorMessages().add(
                         msgUtil.getMessage("listing.invalidUrlFound", fieldName));
             }
@@ -60,9 +65,10 @@ public class UrlReviewer extends PermissionBasedReviewer {
     private void addCriteriaErrorIfNotValid(final CertifiedProductSearchDetails listing,
             final CertificationResult cert, final String input, final String fieldName) {
         if (!StringUtils.isEmpty(input)) {
-            if (ValidationUtils.hasNewline(input) || !ValidationUtils.isWellFormedUrl(input)) {
+            if (validationUtils.hasNewline(input) || !validationUtils.isWellFormedUrl(input)) {
                 addCriterionErrorOrWarningByPermission(listing, cert,
-                        "listing.criteria.invalidUrlFound", fieldName, cert.getNumber());
+                        "listing.criteria.invalidUrlFound", fieldName,
+                        Util.formatCriteriaNumber(cert.getCriterion()));
             }
         }
     }